AT&T's cookbook-networking dependencies
In order for the networking cookbook to properly assign bonding to the network
interfaces, it relies on certain attrs pre-existing.  Usually those attrs are
populated by substructure, which was prior done by manual intervention.  In order
for us to automate the jump node, we want to get networking up FIRST and THEN
have the substructure aspect automated via a cookbook followed after.  In order
to circumvent this problem, for the jumpnode we will populate the attrs without
subs.